requirements.txt
aiofiles==23.2.1
anyio==4.4.0
astroid==3.2.2
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
fonttools==4.53.0
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
ipython==8.25.0
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.2.2
matplotlib-inline==0.1.7
mccabe==0.7.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
prov==2.0.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==2.0.0
Pygments==2.18.0
pylint==3.2.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pyright==1.1.367
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.5
traitlets==5.14.3
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
wcwidth==0.2.13
XlsxWriter==3.2.0
